requirements.txt
flask==2.2.5
flask-login==0.6.2
flask-sqlalchemy==3.0.5
flask-wtf==1.1.1
python-dotenv==1.0.0
requests==2.31.0
gunicorn==21.2.0
openai==1.3.0
python-jose==3.3.0
email-validator==2.1.0
werkzeug==2.2.3
flask-migrate==4.0.5 